Complete 'linear' module
